refactor: 도메인 계층 타입 정리 (라우팅 계층 중복 제거)
PriceChangeDisplay·BreadcrumbItem을 도메인 타입에서 제거하여 UI 관심사 분리,
SkillStat·SkillStatConfig를 skills.ts로 이동하여 유일한 소비처와 함께 배치,
RootLayoutProps 인터페이스 정의로 원시 props 구조 제거
